+++ src/readable_set_ctx_.ts
@@ -1,4 +1,3 @@
-import type { maybe_undefined } from '@ctx-core/function'
 import { Readable, readable } from './readable'
 import { get } from './get'
 import type { StartStopNotifier } from './StartStopNotifier'
@@ -6,7 +5,7 @@ import type { Subscriber } from './Subscriber'
 export function readable_set_ctx_<Val extends unknown = unknown>(
 	initial:Val, readable_fn = readable
 ):readable_set_ctx_T<Val> {
-	let set = undefined as maybe_undefined<_readable_set_ctx_set_T<Val>>
+	let set = undefined as readable_set_ctx$_T<Val>|undefined
 	const store = readable_fn<Val>(initial, (in_set=>{
 		set = in_set
 	}) as StartStopNotifier<Val>) as Readable<Val>
@@ -16,10 +15,10 @@ export function readable_set_ctx_<Val extends unknown = unknown>(
 		set,
 	} as readable_set_ctx_T<Val>
 }
-export type _readable_set_ctx_set_T<Val extends unknown = unknown> = Subscriber<Val>
+export type readable_set_ctx$_T<Val extends unknown = unknown> = Subscriber<Val>
 export interface readable_set_ctx_T<Val extends unknown = unknown> {
 	store:Readable<Val>
-	set:_readable_set_ctx_set_T<Val>
+	set:readable_set_ctx$_T<Val>
 }
 export {
 	readable_set_ctx_ as _readable_set_ctx,