+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ctx-core/store
 
+## 25.1.0
+
+### Minor Changes
+
+- readable*set_a*$_->readable$_set_a_,readable*set_ctx*->$readable$_set_ctx_
+
 ## 25.0.3
 
 ### Patch Changes
@@ -317,7 +323,7 @@
 
 ### Patch Changes
 
-- readable\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\$\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\_C,writable\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\$\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\_C: public get/set \$
+- readable\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\$\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\_C,writable\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\$\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\_C: public get/set \$
 
 ## 24.7.7
 
@@ -373,4 +379,4 @@
 ### Minor Changes
 
 - \_readable_set_ctx_set_type->\_readable_set_ctx_set_T,readable_set_ctx_type->readable_set_ctx_T,\_readable_set_a1_set_type->\_readable_set_a1_set_T,\_readable_set_a1_type->\_readable_set_a1_T
-- feat: \_readable_set_ctx$,_readable_set_a1$
+- feat: \readable*set_ctx\$*,\_readable_set_a1\$